Delay allocating interrupt and fix problems when specifying "ports."
In SAM9X75 board interrupt does not work correctly when allocating
early.  Delay allocating it so that other GPIO interrupts are set.

Re-program PHY DSP registers after powering up as some banks of
registers are reset to default values.

Fix some problems when specifying non-contiguous ports.  Do not setup
SGMII when that port is not enabled as the interrupt will not be
handled when a SFP module is plugged in.