Update KSZ9897 driver to fix a compile issue and allow grouping non-consecutive ports under a network device.
Note the port bitmap used to assign ports uses logical ports, where
the last port is always the host port, even though it can be any
physical port.  For host port being port 6 then the 5 ports before it
use one-to-one mapping but port 7 will be logical port 6. So if the
SGMII port is grouped with port 1 to act as HSR ports then the bitmap
is 0x21.

Currently there are problems using HSR Redbox so please do not use it.